+++ rust/services/inventory/src/tests/inventory_api.rs
@@ -4,7 +4,7 @@ mod tests {
     use anyhow::Result;
     use bson::doc;
     use bson::oid::ObjectId;
-    use db_utils::mongodb::MongoDbAPI;
+    use db_utils::mongodb::api::MongoDbAPI;
     use db_utils::schemas::{workload::Capacity, host::Host};
     use mock_utils::{
         host::create_mock_inventory, mongodb_runner::MongodRunner, nats_message::NatsMessage,
