+++ rust/util_libs/db/src/schemas/user.rs
@@ -3,7 +3,7 @@ use mongodb::options::IndexOptions;
 use serde::{Deserialize, Serialize};
 use anyhow::Result;
 
-use crate::mongodb::{MutMetadata, IntoIndexes};
+use crate::mongodb::traits::{MutMetadata, IntoIndexes};
 use super::metadata::Metadata;
 use super::alias::PubKey;
